Branch - 7
Branch - 7. Breaking the app into components.
