+++ books/src/comps/books.js
@@ -1,0 +1,52 @@
+const books = [
+    {
+      id: 1,
+      image: "https://m.media-amazon.com/images/I/61tJBNx1KsL._AC_UY436_FMwebp_QL65_.jpg",
+      title: "Bruce Lee's Jeet Kune Do",
+      author: "by Sam Fury and Diana Mangoba",
+      desc: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor. Incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,",
+      price: "$14.99"
+    },
+    {
+      id: 2,
+      image: "https://m.media-amazon.com/images/I/71mkvRbiUPL._AC_UY436_FMwebp_QL65_.jpg",
+      title: "Bruce Lee's Fighting Method",
+      author: "by Bruce Lee",
+      desc: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor. Incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,",
+      price: "$26.99"
+    },
+    {
+      id: 3,
+      image: "https://m.media-amazon.com/images/I/91e2UzRiM+L._AC_UY436_FMwebp_QL65_.jpg",
+      title: "Letters of the Dragon",
+      author: "by Bruce Lee",
+      desc: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor. Incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,",
+      price: "$11.99"
+    },
+    {
+        id: 4,
+        image: "https://m.media-amazon.com/images/I/61mZV6fhofL._AC_UY436_FMwebp_QL65_.jpg",
+        title: "Bruce Lee: A Life",
+        author: "by Matthew Polly",
+        desc: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor. Incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,",
+        price: "$14.99"
+    },
+    {
+        id: 5,
+        image: "https://m.media-amazon.com/images/I/61qK5m2EiXL._AC_UY436_FMwebp_QL65_.jpg",
+        title: "Bruce Lee: Jeet Kune Do",
+        author: "by Bruce Lee and John Little",
+        desc: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor. Incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,",
+        price: "$14.99"
+    },
+    {
+        id: 6,
+        image: "https://m.media-amazon.com/images/I/61R3vy2QS0L._AC_UY436_FMwebp_QL65_.jpg",
+        title: "Bruce Lee’s Best Sayings",
+        author: "by Bruce Lee and Joseph Marty",
+        desc: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor. Incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,",
+        price: "$11.99"
+    },
+  ];
+
+  export default books;
